Final Fixes of the Night
Lets just say the sorting and searching is a bit more complicated than I thought. I changed up your merge sort to actually work as it didn't work as it was. I also tried to implement the sort and search but it only returns the full vector in the code. Please help debug!!!!!